More refactoring, moved details-classes to main code, and build them up from the container class
