+++ testcontainers-base/src/main/java/no/vegvesen/ixn/docker/IntermedateCaDetails.java
@@ -1,0 +1,21 @@
+package no.vegvesen.ixn.docker;
+
+import java.nio.file.Path;
+
+class IntermedateCaDetails {
+    private final Path singleCertOnHost;
+    private final Path chainCertOnHost;
+
+    public IntermedateCaDetails(Path singleCertOnHost, Path chainCertOnHost) {
+        this.singleCertOnHost = singleCertOnHost;
+        this.chainCertOnHost = chainCertOnHost;
+    }
+
+    public Path getSingleCertOnHost() {
+        return singleCertOnHost;
+    }
+
+    public Path getChainCertOnHost() {
+        return chainCertOnHost;
+    }
+}
