+++ testcontainers-base/src/main/java/no/vegvesen/ixn/docker/RootCaDetails.java
@@ -1,0 +1,22 @@
+package no.vegvesen.ixn.docker;
+
+import java.nio.file.Path;
+
+public class RootCaDetails {
+
+    private final Path caKeyOnHost;
+    private final Path caCertOnHost;
+
+    public RootCaDetails(Path caKeyOnHost, Path caCertOnHost) {
+        this.caKeyOnHost = caKeyOnHost;
+        this.caCertOnHost = caCertOnHost;
+    }
+
+    public Path getCaKeyOnHost() {
+        return caKeyOnHost;
+    }
+
+    public Path getCaCertOnHost() {
+        return caCertOnHost;
+    }
+}
